fix eigenda-operator-state unit tests
