+++ subgraphs/eigenda-operator-state/src/operator-registration-status.ts
@@ -2,10 +2,10 @@ import { BigInt, Bytes, log } from "@graphprotocol/graph-ts"
 import {
   OperatorRegistered as OperatorRegisteredEvent,
   OperatorDeregistered as OperatorDeregisteredEvent
-} from "../generated/BLSRegistryCoordinatorWithIndices_Operator/BLSRegistryCoordinatorWithIndices"
-import { NewPubkeyRegistration as NewPubkeyRegistrationEvent } from "../generated/BLSPubkeyCompendium/BLSPubkeyCompendium"
+} from "../generated/RegistryCoordinator_Operator/RegistryCoordinator"
+import { NewPubkeyRegistration as NewPubkeyRegistrationEvent } from "../generated/BLSApkRegistry/BLSApkRegistry"
 import { Operator } from "../generated/schema"
-import { BLSPubkeyCompendium } from "../generated/BLSPubkeyCompendium/BLSPubkeyCompendium"
+import { BLSApkRegistry } from "../generated/BLSApkRegistry/BLSApkRegistry"
 
 export function handleOperatorDeregistered(event: OperatorDeregisteredEvent) : void {
   let entity = Operator.load(event.params.operatorId)
